+++ src/goc_mpc_cpp/goc_mpc/graph_ordering_mpc_problem.hpp
@@ -33,6 +33,6 @@ struct GraphOrderingProblem {
 
 GraphOrderingProblem build_graph_ordering_problem(
 	const py::array_t<double>& waypoints,
-	const py::array_t<double>& graph,
+	const py::array_t<unsigned int>& graph,
 	const py::array_t<double>& x0_np,
 	const py::array_t<double>& v0_np);
